% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddt.R
\docType{data}
\name{ddt}
\alias{ddt}
\title{DDT Contamination Data}
\format{
## `ddt`
A data frame with multiple rows and 6 columns:
\describe{
  \item{RIVER}{River location where the sample was collected}
  \item{MILE}{Mile marker of the collection site along the river}
  \item{SPECIES}{Fish species name}
  \item{LENGTH}{Fish length (cm)}
  \item{WEIGHT}{Fish weight (grams)}
  \item{DDT}{Measured DDT concentration in fish tissue (ppm)}
}
}
\source{
https://www.amazon.com/Statistics-Engineering-Sciences-William-Mendenhall/dp/1498728855
}
\usage{
ddt
}
\description{
A dataset containing measurements related to DDT contamination in fish samples, including location, species, and physical attributes.
}
\keyword{datasets}
