% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendiff.R
\name{dendiff}
\alias{dendiff}
\title{Density Difference Function for Beta HDI with Graph Output}
\arguments{
\item{a1}{Numeric. Left-tail exclusion probability.}

\item{shape1}{Numeric. First shape parameter of the Beta distribution.}

\item{shape2}{Numeric. Second shape parameter of the Beta distribution.}

\item{alpha}{Numeric. Total tail probability for the interval.}
}
\value{
Numeric value representing the difference in density.
Also displays a plot of density difference vs. a1 with intersecting lines.
}
\description{
Calculates the density difference between two Beta quantiles given a tail probability.
Also displays a diagnostic plot of the density difference curve across possible `a1` values,
aiding in visual root-finding to identify HDIs.
}
\examples{
dendiff(a1 = 0.02, shape1 = 3, shape2 = 10, alpha = 0.05)
}
